Generalize tree and separate into modules
