+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "red_and_black"
+name = "baltree"
 version = "0.1.0"
 edition = "2021"
