+++ src/red_black_tree.rs
@@ -1,0 +1,9 @@
+use crate::balancing_tree::{BalancingTree, GenericTree, Node, Tree};
+
+pub struct RedBlackTree {
+    pub tree: BalancingTree,
+}
+
+impl RedBlackTree {
+    // RBT-specific methods
+}
